requirements.txt
streamlit
openai
sympy
pydantic
python-dotenv